mamoru_agent_api/pyproject.toml
[project]
name = "mamoru-agent-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anyio>=4.8.0",
    "fastapi>=0.115.8",
    "pydantic>=2.10.6",
    "pydantic-ai>=0.0.23",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
]
